Add make_anvil_same_annotator
Script to change the "coder" field in a directory of anvil files to a
single annotator. Useful to create a temporary set of annotations with
the same annotator, which allows one to run anvil_to_frame.py to create
a table of all the videos ground truth annotations per second. This
table is then useful to run statistics on ground-truth phase lengths.