Add example_phase_factors.csv
This is a csv that shows an example of a file to use as "FACTORS" for
the R statistic scripts. It contains structure of (name,full_name). Its
use is to allow for the statistics and plots generated with the
statistic scripts to display the "human readable" name (eg instead of
showing muc_incis, it shows Mucosotomy) and in the correct order. All
the R scripts will base the steps' order based on the order in the file.
So you see in this example, submuc_inject is followed by muc_incis, etc.
This is the temporal order those steps typically happen in the
operation, so will then be displayed in an order that makes sense in the
plots rather than in simply lexigraphical order.