Add overall_phase_stats.R
overall_phase_stats.R is an R script that will generate statistics
on phase durations across videos that it will save in a TSV file and
optionally plot a boxplot, in either log or not log scale. It was used
to generate statistics for the POEM paper.

It takes:

1. CSV file that contains info under header (video,frame,annotator)
(frame is equivalent to second for this purpose). This is generated with
anvil_to_frame.csv.

2. CSV file with (name,full_name) header that contains mapping from name
the phase is represented as in the model to the human readable name you
want displayed on a plot.

It outputs

1. TSV with spread statistics for how each phase's duration varies across videos.
This the phase's duration over an entire video, *not* in blocks. Outputs
mean, standard deviation, min, 1st quartile, median, 3rd quartile, max,
MAD.

2. Optional boxplot. This can optionally be on a logarithmic scale.