Add anvil annotation examples for multiple videos
Each anvil annotation is an xml file that contains the annotation for a
single video. The filename itself is not important for any of the
scripts. What is important is the contents, including having each
same annotator be named the same thing in the xml `coder` field.
For example, my name is "Thomas Ward" and I always put "tmw" as my
annotator name. If instead I sometimes used "tmw" then "TW" in other
times, it would be counted as different annotators rather than the same.