Add model_metrics.R
An R script that will, given the output from make_model_results.tsv
and a csv that holds (short_phase_name,long_phase_name), will generate
statistics on the model's performance, on an overall, and per-video
basis, including:
	1. All raw results (combined_raw.tsv or video_NN_raw.tsv)
	2. tidied confusion matrix (combined_confusion.tsv or
	   video_NN_confusion.tsv)
	3. Table with per-phase metrics and overall (weighted and
	   unweighted metrics) called (combined_perclass.tsv or
	   video_NN_perclass.tsv)
	4. Confusion matrix image for recall (combined_recall.png or
	   video_NN_recall.png)
	5. Confusion matrix image for precision (combined_precision.png or
	   video_NN_precision.png)

The combined_perclass.tsv table is identical to the table in the POEM
paper. Raw results holds overall model accuracy and 95% CI as well in
addition to all the stats. confusion.tsv holds the correct numbers from
which to calculate confusion matrixes.