Add durations.csv
This is an example file that shows how to write a csv of durations on
which to then analyse for the model. It has two columns (start, end). In
order to say "until the end" when you don't know the maximum length of a
duration, use "Inf" which in R stands for infinity.