Add per_block_duration_accuracy.R
R script that will calculate the model's accuracy for blocks of
different durations. A block is a contiguous occurrence of a certain
ground truth before it transitions to a different ground truth. Instead
of analyzing on a per-phase basis, we instead look at how it does with
phase blocks of different durations. This helped identify issues with
extremely short duration blocks and accuracy.

To specify durations, can write a csv file with two columns, (start, end).
To specify the end of a video, use "Inf" which is R for infinity. Can
see an example of the format in example/durations.csv.